requirement.txt
Django>=3.1.1<4.0.0
djangorestframework>=3.11.1<4.0.0
django-filter>=2.4.0<3.0.0
Pillow>=7.2.0<8.0.0
django-cors-headers>=3.5.0<4.0.0
flake8>=3.8.2<4.0.0
Faker>=4.4.0<5.0.0